Debugged spaced stacked view, added user param
User parameter "centred" for whether genotypes should be centred or not (stacked) has been changed to "genotype.position", where genotypes are specified to be "stack", "centre" or "space".
Debugged spaced view, mostly an issue of the ancestor not having any cellular prevalence when it's replaced, and dealing with placing its descendants accordingly.