Increase width for low cellular prevalence
Created a function to get the siblings of each node in the tree.
Created a function to get the effective cellular prevalence of a genotype at a particular time point - this is affected by siblings whose lineage is all low cellular prevalence (genotype will be shrunk), and is also affected by genotypes that themselves have low cellular prevalence (genotype will be increased to threshold value).